Fix #827: Requests with unknown HTTP method result in uncaught exceptions #827
