+++ http4k-core/build.gradle.kts
@@ -14,6 +14,7 @@ dependencies {
     testApi(project(":http4k-server-apache"))
     testApi("org.webjars:swagger-ui:4.11.1") // leave hardcoded - tests
     testApi("com.launchdarkly:okhttp-eventsource:_")
+    testApi("org.apache.httpcomponents:httpcore:_")
     testApi("org.slf4j:slf4j-nop:_")
 
     testImplementation(Testing.junit.jupiter.params)
