+++ http4k-server/apache4/src/main/kotlin/org/http4k/server/Apache4Server.kt
@@ -9,7 +9,7 @@ import org.apache.http.impl.bootstrap.HttpServer
 import org.apache.http.impl.bootstrap.ServerBootstrap
 import org.apache.http.impl.io.EmptyInputStream
 import org.apache.http.protocol.HttpContext
-import org.apache.http.protocol.HttpCoreContext
+import org.apache.http.protocol.HttpCoreContext.HTTP_CONNECTION
 import org.apache.http.protocol.HttpRequestHandler
 import org.http4k.core.Headers
 import org.http4k.core.HttpHandler
@@ -17,6 +17,8 @@ import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.RequestSource
 import org.http4k.core.Response
+import org.http4k.core.Status
+import org.http4k.core.Status.Companion.NOT_IMPLEMENTED
 import org.http4k.core.safeLong
 import org.http4k.core.then
 import org.http4k.filter.ServerFilters
@@ -33,20 +35,26 @@ class Http4kApache4RequestHandler(handler: HttpHandler) : HttpRequestHandler {
 
     private val safeHandler = ServerFilters.CatchAll().then(handler)
 
-    override fun handle(request: ApacheRequest, response: ApacheResponse, context: HttpContext) {
-        safeHandler(request.asHttp4kRequest(context)).into(response)
-    }
+    override fun handle(request: ApacheRequest, response: ApacheResponse, context: HttpContext) =
+        (request.asHttp4kRequest(context)?.let(safeHandler) ?: Response(NOT_IMPLEMENTED)).into(response)
 
-    private fun ApacheRequest.asHttp4kRequest(context: HttpContext): Request {
-        val connection = context.getAttribute(HttpCoreContext.HTTP_CONNECTION) as HttpInetConnection
-        return Request(Method.valueOf(requestLine.method), requestLine.uri)
-            .headers(allHeaders.toHttp4kHeaders()).let {
-                when (this) {
-                    is HttpEntityEnclosingRequest -> it.body(entity.content, getFirstHeader("Content-Length")?.value.safeLong())
-                    else -> it.body(EmptyInputStream.INSTANCE, 0)
+    private fun ApacheRequest.asHttp4kRequest(context: HttpContext): Request? {
+        val connection = context.getAttribute(HTTP_CONNECTION) as HttpInetConnection
+        return runCatching {
+            Request(Method.valueOf(requestLine.method), requestLine.uri)
+                .headers(allHeaders.toHttp4kHeaders()).let {
+                    when (this) {
+                        is HttpEntityEnclosingRequest -> it.body(
+                            entity.content,
+                            getFirstHeader("Content-Length")?.value.safeLong()
+                        )
+
+                        else -> it.body(EmptyInputStream.INSTANCE, 0)
+                    }
                 }
-            }
-            .source(RequestSource(connection.remoteAddress.hostAddress, connection.remotePort))
+                .source(RequestSource(connection.remoteAddress.hostAddress, connection.remotePort))
+
+        }.getOrNull()
     }
 
     private val headersThatApacheInterceptorSets = setOf("Transfer-Encoding", "Content-Length")
@@ -55,7 +63,8 @@ class Http4kApache4RequestHandler(handler: HttpHandler) : HttpRequestHandler {
         with(response) {
             setStatusCode(status.code)
             setReasonPhrase(status.description)
-            headers.filter { !headersThatApacheInterceptorSets.contains(it.first) }.forEach { (key, value) -> addHeader(key, value) }
+            headers.filter { !headersThatApacheInterceptorSets.contains(it.first) }
+                .forEach { (key, value) -> addHeader(key, value) }
             entity = InputStreamEntity(body.stream, body.length ?: -1L)
         }
     }
@@ -69,7 +78,7 @@ class Apache4Server(val port: Int = 8000, override val stopMode: StopMode, val a
     constructor(port: Int = 8000) : this(port, null)
 
     init {
-        if(stopMode != StopMode.Immediate){
+        if (stopMode != StopMode.Immediate) {
             throw ServerConfig.UnsupportedStopMode(stopMode)
         }
     }
@@ -80,12 +89,14 @@ class Apache4Server(val port: Int = 8000, override val stopMode: StopMode, val a
         init {
             val bootstrap = ServerBootstrap.bootstrap()
                 .setListenerPort(port)
-                .setSocketConfig(SocketConfig.custom()
-                    .setTcpNoDelay(true)
-                    .setSoKeepAlive(true)
-                    .setSoReuseAddress(true)
-                    .setBacklogSize(1000)
-                    .build())
+                .setSocketConfig(
+                    SocketConfig.custom()
+                        .setTcpNoDelay(true)
+                        .setSoKeepAlive(true)
+                        .setSoReuseAddress(true)
+                        .setBacklogSize(1000)
+                        .build()
+                )
                 .registerHandler("*", Http4kApache4RequestHandler(http))
 
             if (address != null)
