+++ http4k-server/ktorcio/src/main/kotlin/org/http4k/server/KtorCIO.kt
@@ -26,7 +26,7 @@ import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.RequestSource
 import org.http4k.core.Response
-import org.http4k.lens.Header
+import org.http4k.core.Status.Companion.NOT_IMPLEMENTED
 import org.http4k.lens.Header.CONTENT_TYPE
 import org.http4k.server.ServerConfig.StopMode.Immediate
 import java.util.concurrent.TimeUnit.SECONDS
@@ -47,7 +47,7 @@ class KtorCIO(val port: Int = 8000, override val stopMode: ServerConfig.StopMode
             install(createApplicationPlugin(name = "http4k") {
                 onCall {
                     withContext(Default) {
-                        it.response.fromHttp4K(http(it.request.asHttp4k()))
+                        it.response.fromHttp4K(it.request.asHttp4k()?.let(http) ?: Response(NOT_IMPLEMENTED))
                     }
                 }
             })
@@ -65,10 +65,12 @@ class KtorCIO(val port: Int = 8000, override val stopMode: ServerConfig.StopMode
     }
 }
 
-fun ApplicationRequest.asHttp4k() = Request(Method.valueOf(httpMethod.value), uri)
-    .headers(headers.toHttp4kHeaders())
-    .body(receiveChannel().toInputStream(), header("Content-Length")?.toLong())
-    .source(RequestSource(origin.remoteHost, scheme = origin.scheme)) // origin.remotePort does not exist for Ktor
+fun ApplicationRequest.asHttp4k() = runCatching {
+    Request(Method.valueOf(httpMethod.value), uri)
+        .headers(headers.toHttp4kHeaders())
+        .body(receiveChannel().toInputStream(), header("Content-Length")?.toLong())
+        .source(RequestSource(origin.remoteHost, scheme = origin.scheme)) // origin.remotePort does not exist for Ktor
+}.getOrNull()
 
 suspend fun ApplicationResponse.fromHttp4K(response: Response) {
     status(HttpStatusCode.fromValue(response.status.code))
