+++ http4k-server/ratpack/src/main/kotlin/org/http4k/server/Ratpack.kt
@@ -1,14 +1,9 @@
 package org.http4k.server
 
 import org.http4k.core.HttpHandler
-import org.http4k.core.Method
-import org.http4k.core.Request
-import org.http4k.core.RequestSource
-import org.http4k.core.Response
 import org.http4k.server.ServerConfig.StopMode
-import ratpack.handling.Context
-import ratpack.handling.Handler
-import ratpack.http.TypedData
+import org.http4k.server.ServerConfig.StopMode.Graceful
+import org.http4k.server.ServerConfig.UnsupportedStopMode
 import ratpack.server.RatpackServer
 import ratpack.server.RatpackServerSpec
 import ratpack.server.ServerConfig.builder
@@ -18,7 +13,7 @@ class Ratpack(port: Int = 8000, stopMode: StopMode) : ServerConfig {
 
     init {
         when (stopMode) {
-            is StopMode.Graceful -> throw ServerConfig.UnsupportedStopMode(stopMode)
+            is Graceful -> throw UnsupportedStopMode(stopMode)
             else -> {}
         }
     }
@@ -32,42 +27,11 @@ class Ratpack(port: Int = 8000, stopMode: StopMode) : ServerConfig {
         }
 
         return object : Http4kServer {
-            override fun start(): Http4kServer = apply { server.start() }
+            override fun start() = apply { server.start() }
 
-            override fun stop(): Http4kServer = apply { server.stop() }
+            override fun stop() = apply { server.stop() }
 
-            override fun port(): Int = server.bindPort
+            override fun port() = server.bindPort
         }
-    }
-}
-
-class RatpackHttp4kHandler(private val httpHandler: HttpHandler) : Handler {
-    override fun handle(context: Context) {
-        context.request.body.then { data ->
-            val request = context.toHttp4kRequest(data)
-            val response = httpHandler(request)
-            response.pushTo(context)
-        }
-    }
-
-    private fun Context.toHttp4kRequest(data: TypedData) = Request(Method.valueOf(request.method.name), request.rawUri)
-        .let {
-            request.headers.names.fold(it, { acc, nextHeaderName ->
-                request.headers.getAll(nextHeaderName)
-                    .fold(acc, { vAcc, nextHeaderValue ->
-                        vAcc.header(nextHeaderName, nextHeaderValue)
-                    })
-            })
-        }
-        .body(data.inputStream, request.headers.get("content-length")?.toLongOrNull())
-        .source(RequestSource(request.remoteAddress.host, request.remoteAddress.port))
-
-    private fun Response.pushTo(context: Context) {
-        headers.groupBy { it.first }
-            .forEach { (name, values) ->
-                context.response.headers.set(name, values.mapNotNull { it.second })
-            }
-        context.response.status(status.code)
-        context.response.send(body.payload.array())
     }
 }
